[DCGEO-145] Commonality with v1 samhsa
