+++ apps/bfd-server/bfd-server-war/src/test/java/gov/cms/bfd/server/war/commons/AbstractSamhsaMatcherTest.java
@@ -1,0 +1,17 @@
+package gov.cms.bfd.server.war.commons;
+
+import static org.junit.Assert.assertEquals;
+
+import java.util.List;
+import org.junit.Test;
+
+public class AbstractSamhsaMatcherTest {
+
+  @Test
+  public void shouldReturnColumnValues() {
+    List<String> expected = List.of("1", "2", "3", "4");
+    List<String> actual = AbstractSamhsaMatcher.resourceCsvColumnToList("test_file.csv", "columnB");
+
+    assertEquals(expected, actual);
+  }
+}
